[desktop] Speed up reconciliation by doing an upfront directory listing
